Fix executor starting late and keeping the processes running
The issue was a race condition. A world was created, but before it could be
registered as @world in the BenchmarkHelper, its process received a SIGTERM.
Upon that we tried to terminate the world, if present. With this we set the
termination to happen when the world is finally set, so it doesn't slip through.